R49/command_Factory/CommandFactory.ts
// Define the Task interface
interface Task {
    name: string;
    status: 'pending' | 'in_progress' | 'completed';
    dueDate: Date;
  }
  
  // Define the CommandType enum
  enum CommandType {
    CreateTask,
    UpdateTaskStatus,
    GetTaskStatus,
  }
  
  // Define the Command interface
  interface Command {
    execute(): void;
  }
  
  // Define the CreateTaskCommand class
  class CreateTaskCommand implements Command {
    private taskName: string;
    private dueDate: Date;
  
    constructor(taskName: string, dueDate: Date) {
      this.taskName = taskName;
      this.dueDate = dueDate;
    }
  
    execute(): void {
      // Add the task to the project management system
      addTask(this.taskName, this.dueDate);
    }
  }
  
  // Define the UpdateTaskStatusCommand class
  class UpdateTaskStatusCommand implements Command {
    private taskName: string;
    private status: 'pending' | 'in_progress' | 'completed';
  
    constructor(taskName: string, status: 'pending' | 'in_progress' | 'completed') {
      this.taskName = taskName;
      this.status = status;
    }
  
    execute(): void {
      // Update the task status in the project management system
      updateTaskStatus(this.taskName, this.status);
    }
  }
  
  // Define the GetTaskStatusCommand class
  class GetTaskStatusCommand implements Command {
    private taskName: string;
  
    constructor(taskName: string) {
      this.taskName = taskName;
    }
  
    execute(): void {
      // Get the task status from the project management system
      getTaskStatus(this.taskName);
    }
  }
  
  // Define the CommandFactory class
  class CommandFactory {
    createCommand(command: string): Command | null {
      // Use a state machine to parse the command
      const words = command.split(' ');
      let commandType: CommandType | null = null;
      let taskName: string | null = null;
      let dueDate: Date | null = null;
      let status: 'pending' | 'in_progress' | 'completed' | null = null;
  
      for (const word of words) {
        if (word === 'Create') {
          commandType = CommandType.CreateTask;
        } else if (word === 'Mark') {
          commandType = CommandType.UpdateTaskStatus;
        } else if (word === 'What') {
          commandType = CommandType.GetTaskStatus;
        } else if (word.startsWith("'") && word.endsWith("'")) {
          taskName = word.slice(1, -1);
        } else if (word === 'with' && words[words.indexOf(word) + 1] === 'a' && words[words.indexOf(word) + 2] === 'due' && words[words.indexOf(word) + 3] === 'date' && words[words.indexOf(word) + 4] === 'of') {
          dueDate = new Date(words[words.indexOf(word) + 5]);
        } else if (word === 'as') {
          status = words[words.indexOf(word) + 1].toLowerCase() as 'pending' | 'in_progress' | 'completed';
        }
      }
  
      if (commandType === null || taskName === null) {
        return null;
      }
  
      switch (commandType) {
        case CommandType.CreateTask:
          return new CreateTaskCommand(taskName, dueDate!);
        case CommandType.UpdateTaskStatus:
          return new UpdateTaskStatusCommand(taskName, status!);
        case CommandType.GetTaskStatus:
          return new GetTaskStatusCommand(taskName);
        default:
          return null;
      }
    }
  }
  
  // Define the addTask function
  function addTask(taskName: string, dueDate: Date): void {
    // Add the task to the project management system
    tasks.push({
      name: taskName,
      status: 'pending',
      dueDate,
    });
  
    // Respond with a confirmation message
    console.log(`Task '${taskName}' created with a due date of ${dueDate.toLocaleDateString()}.`);
  }
  
  // Define the updateTaskStatus function
  function updateTaskStatus(taskName: string, status: 'pending' | 'in_progress' | 'completed'): void {
    // Find the task to update
    const taskToUpdate = tasks.find((task) => task.name === taskName);
  
    // Check if the task exists
    if (taskToUpdate) {
      // Update the task status
      taskToUpdate.status = status;
  
      // Respond with a confirmation message
      console.log(`Task '${taskName}' marked as ${status}.`);
    } else {
      // Respond with an error message if the task does not exist
      console.log(`Task '${taskName}' not found.`);
    }
  }
  
  // Define the getTaskStatus function
  function getTaskStatus(taskName: string): void {
    // Find the task to retrieve the status for
    const task = tasks.find((task) => task.name === taskName);
  
    // Check if the task exists
    if (task) {
      // Respond with the task status
      console.log(`The status of task '${taskName}' is ${task.status}.`);
    } else {
      // Respond with an error message if the task does not exist
      console.log(`Task '${taskName}' not found.`);
    }
  }
  
  // Initialize an array to simulate a project management system
  let tasks: Task[] = [];
  
  // Test the conversational assistant
  const commandFactory = new CommandFactory();
  const createTaskCommand = commandFactory.createCommand("Create a task called 'Design Homepage' with a due date of October 5, 2024");
  const updateTaskStatusCommand = commandFactory.createCommand("Mark the 'Design Homepage' task as completed");
  const getTaskStatusCommand = commandFactory.createCommand("What is the status of the 'Design Homepage' task?");
  
  if (createTaskCommand) {
    createTaskCommand.execute();
  }
  
  if (updateTaskStatusCommand) {
    updateTaskStatusCommand.execute();
  }
  
  if (getTaskStatusCommand) {
    getTaskStatusCommand.execute();
  }